+++ db/migrate/20170720172049_change_remote_id_on_activities_to_text.rb
@@ -1,9 +1,0 @@
-class ChangeRemoteIdOnActivitiesToText < ActiveRecord::Migration
-  def up
-    change_column :activities, :remote_id, :text
-  end
-
-  def down
-    change_column :activities, :remote_id, :string
-  end
-end